Use historyMiddleware factory
