+++ src/redux/middlewares.js
@@ -2,7 +2,7 @@
 import * as _ from 'lodash-es';
 import * as Sentry from '@sentry/react';
 
-import { LOCATION_CHANGE, locationPush } from '../services/nouter/redux';
+import { LOCATION_CHANGE, LOCATION_PUSH, locationPush } from '../services/nouter/redux';
 import { getPost } from '../services/api';
 import { getToken, setToken } from '../services/auth';
 import { Connection } from '../services/realtime';
@@ -33,6 +33,7 @@ import {
   doneEditingAndDeleteDraft,
   initializeDrafts,
 } from '../services/drafts';
+import { withQuery } from '../services/nouter/utils';
 import * as ActionCreators from './action-creators';
 import * as ActionTypes from './action-types';
 import {
@@ -1101,8 +1102,7 @@ export const reloadFeedMiddleware = (store) => (next) => (action) => {
         userName === action.request.username)
     ) {
       // Re-request this page
-      console.log('Reloading feed');
-      store.dispatch(locationPush(`/${userName}`));
+      store.dispatch(ActionCreators.getUserFeed(userName));
     }
   }
 
@@ -1189,6 +1189,38 @@ export function undoMiddleware(store) {
         rtHandlers['post:restore'](action.payload);
       }
     }
+  };
+}
+
+/**
+ * The `$Factory` is a special suffix for the factory that allows to pass
+ * additional arguments to the middleware.
+ */
+export function historyMiddleware$Factory({ history }) {
+  return (store) => {
+    function onLocationChange({ location }) {
+      store.dispatch({
+        type: LOCATION_CHANGE,
+        payload: withQuery({
+          pathname: location.pathname,
+          search: location.search,
+          hash: location.hash,
+        }),
+      });
+    }
+    history.listen(onLocationChange);
+    // Initial dispatch
+    setTimeout(() => onLocationChange({ location: history.location }), 0);
+
+    return (next) => (action) => {
+      // A patch for calling history API via Redux actions. It is necessary for
+      // middlewares that want to manage history.
+      if (action.type === LOCATION_PUSH) {
+        const { to, replace } = action.payload;
+        history[replace ? 'replace' : 'push'](to);
+      }
+      return next(action);
+    };
   };
 }
